add bin build function
